+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 4.2.2
+
+### Enhancements
+- Add FDS and Consulting Referral email QuickLink.  
+- Upgrade Font Awesome to 7.2.0  
+
+### Tweaks
+- Correct response handling for internal messaging.  
+- Improve background keepAlive function.   
+
 # 4.2.1
 
 ### Fixes